Refactor CardDisplayOptions
