+++ src/components/sections/CardDisplayOptions/getExternalPropertiesOptionGetter.js
