+++ src/components/sections/SearchOptionsForExternalData.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { compose } from 'recompose';
 import * as Fields from '../../ui/components/Fields';
 import ValidatedField from '../Form/ValidatedField';
@@ -7,15 +8,13 @@ import withMapFormToProps from '../enhancers/withMapFormToProps';
 import withExternalData from '../enhancers/withExternalData';
 import Section from './Section';
 import Row from './Row';
-import withExternalDataPropertyOptions from './SortOptionsForExternalData/withExternalDataPropertyOptions';
+import withVariableOptionsFromExternalData from '../enhancers/withVariableOptionsFromExternalData';
 
 const SearchOptions = ({
-  maxAdditionalDisplayProperties,
-  additionalPropertiesOptionGetter,
   variableOptions,
-  ...props
+  disabled,
 }) => (
-  <Section group contentId="guidance.editor.SearchOptions" {...props}>
+  <Section group contentId="guidance.editor.SearchOptions" disabled={disabled}>
     <Row>
       <h3>Search Options</h3>
       <p>
@@ -62,12 +61,17 @@ const SearchOptions = ({
   </Section>
 );
 
+SearchOptions.propTypes = {
+  variableOptions: PropTypes.array.isRequired,
+  disabled: PropTypes.bool.isRequired,
+};
+
 export { SearchOptions };
 
 export default compose(
   withMapFormToProps(['dataSource']),
   withDisabledAssetRequired,
   withExternalData,
-  withExternalDataPropertyOptions,
+  withVariableOptionsFromExternalData,
 )(SearchOptions);
